Integer cast changed to Long.

+++ src/model/Instruction.java
@@ -2,11 +2,11 @@ package model;
 
 public class Instruction {
 
-    public int countOfPlayers, startMoney, countOfTaxSquares, priceOfTaxSquares, priceOfGoSquare;
+    public long countOfPlayers, startMoney, countOfTaxSquares, priceOfTaxSquares, priceOfGoSquare;
 
     private static Instruction instance;
 
-    public Instruction(int countOfPlayers, int startMoney, int countOfTaxSquares, int priceOfTaxSquares, int priceOfGoSquare) {
+    public Instruction(long countOfPlayers, long startMoney, long countOfTaxSquares, long priceOfTaxSquares, long priceOfGoSquare) {
         this.countOfPlayers = countOfPlayers;
         this.startMoney = startMoney;
         this.countOfTaxSquares = countOfTaxSquares;
@@ -18,8 +18,8 @@ public class Instruction {
         return instance;
     }
 
-    public static Instruction getInstance(int countOfPlayers, int startMoney, int countOfTaxSquares,
-                                          int priceOfTaxSquares, int priceOfGoSquare) {
+    public static Instruction getInstance(long countOfPlayers, long startMoney, long countOfTaxSquares,
+                                          long priceOfTaxSquares, long priceOfGoSquare) {
         if (instance == null) {
             instance = new Instruction(countOfPlayers, startMoney, countOfTaxSquares,
                     priceOfTaxSquares, priceOfGoSquare);
